+++ src/main/kotlin/org/sonso/bludceapi/client/OcrServiceClient.kt
@@ -1,7 +1,7 @@
 package org.sonso.bludceapi.client
 
 import org.sonso.bludceapi.config.FeignMultipartSupportConfig
-import org.sonso.bludceapi.dto.response.ocr.OcrServiceClientResponse
+import org.sonso.bludceapi.dto.response.OcrServiceClientResponse
 import org.springframework.cloud.openfeign.FeignClient
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestPart
@@ -9,7 +9,7 @@ import org.springframework.web.multipart.MultipartFile
 
 @FeignClient(
     name = "ocrClient",
-    url = "https://api.ocr.space",
+    url = "\${app.ocr.url}",
     configuration = [FeignMultipartSupportConfig::class]
 )
 interface OcrServiceClient {
