--- src/main/kotlin/org/sonso/bludceapi/dto/response/ocr/OcrServiceClientResponse.kt
@@ -1,20 +1,0 @@
-package org.sonso.bludceapi.dto.response.ocr
-
-import com.fasterxml.jackson.annotation.JsonProperty
-
-data class OcrServiceClientResponse(
-    @JsonProperty("ParsedResults")
-    val parsedResults: List<ParsedResultResponse>,
-
-    @JsonProperty("OCRExitCode")
-    val ocrExitCode: Int,
-
-    @JsonProperty("IsErroredOnProcessing")
-    val isErroredOnProcessing: Boolean,
-
-    @JsonProperty("ProcessingTimeInMilliseconds")
-    val processingTimeInMilliseconds: String,
-
-    @JsonProperty("SearchablePDFURL")
-    val searchablePDFURL: String
-)
