+++ src/main/resources/application.yml
@@ -13,9 +13,10 @@ spring:
     url: jdbc:postgresql://localhost:5700/bludce   # idea
     driver-class-name: org.postgresql.Driver
 
-ocr:
-  api:
-    key: K88899112888957
+app:
+  ocr:
+    api-key: K88899112888957
+    url: https://api.ocr.space
 
 server:
   port: 7200
